integrate old site cold into template
this adds all the base site code that was in the old. no changes to
functionality, just integrating into the grunt build and template.

for the build, had to comment out imagemin because there seems to
be a bug with that. so for now, it just copies over the images/*
and uses the existing.